Restructure the code
